feat: Created plugin_utils folder and moved misc files there
